Update e2e tests
