--- pipelines/tests/conftest.py
@@ -1,29 +1,0 @@
-# Copyright 2022 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import pytest
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--enable_caching",
-        type=str,
-        help="Whether to enable or disable caching for all pipeline steps",
-        default=None,
-    )
-
-
-@pytest.fixture(scope="session")
-def enable_caching(pytestconfig):
-    return pytestconfig.getoption("enable_caching")
